requirements.txt
Pillow~=9.1.1
loguru~=0.6.0